+++ package.json
@@ -2,14 +2,14 @@
   "name": "repond-movers",
   "version": "0.7.0",
   "description": "movers for repond",
-  "main": "src/index.ts",
-  "types": "src/index.ts",
-  "files": [
+  "_main": "src/index.ts",
+  "_types": "src/index.ts",
+  "_files": [
     "src"
   ],
-  "_main": "dist/index.js",
-  "_types": "dist/index.d.ts",
-  "_files": [
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "files": [
     "dist"
   ],
   "scripts": {
@@ -30,14 +30,12 @@
   "license": "MIT",
   "devDependencies": {
     "@types/react": "^17.0.28",
-    "chootils": "^0.3.8",
+    "chootils": "^0.3.9",
     "react": "^17.0.2",
-    "typescript": "^4.4.3"
+    "typescript": "^4.4.3",
+    "repond": "^0.10.0"
   },
   "prettier": {
     "printWidth": 120
-  },
-  "dependencies": {
-    "repond": "^0.9.3"
   }
 }
