+++ src/types.ts
@@ -1,5 +1,13 @@
 // concepto-movers
 
+import { AllState, AllRefs } from "repond";
+
+export type MoverType = "1d" | "2d" | "3d" | "multi";
+export type ItemType = keyof AllState & keyof AllRefs;
+export type ItemState<T_ItemType extends ItemType> = AllState[T_ItemType][keyof AllState[T_ItemType]];
+export type ItemRefs<T_ItemType extends ItemType> = AllRefs[T_ItemType][keyof AllRefs[T_ItemType]];
+export type StateNameProperty<T_ItemType extends ItemType> = keyof ItemState<T_ItemType> & string;
+
 // might use manual instead of dragging
 export type MoveMode = "spring" | "slide" | "drag" | "push";
 export type PhysicsOptions = {
@@ -23,12 +31,7 @@ export type OnePhysicsConfig = {
 export type TempStateNameProperty = string;
 
 type OnePhysicsConfigOptions = Partial<OnePhysicsConfig> &
-  (
-    | { mass: number }
-    | { stiffness: number }
-    | { damping: number }
-    | { friction: number }
-  );
+  ({ mass: number } | { stiffness: number } | { damping: number } | { friction: number });
 
 type MultipleConfigsOption = Record<string, OnePhysicsConfigOptions> & {
   mass?: undefined;
@@ -38,10 +41,7 @@ type MultipleConfigsOption = Record<string, OnePhysicsConfigOptions> & {
   friction?: undefined;
 };
 
-export type PhysicsConfig =
-  | undefined
-  | OnePhysicsConfigOptions
-  | MultipleConfigsOption;
+export type PhysicsConfig = undefined | OnePhysicsConfigOptions | MultipleConfigsOption;
 
 export type DefinedPhysicsConfig = Record<string, OnePhysicsConfig>;
 
@@ -75,3 +75,12 @@ export type AnyMoverStateNames = {
 //   moveMode?: StateNameProperty<T_ItemType>;
 //   physicsConfigs?: StateNameProperty<T_ItemType>;
 // };
+
+export type RunMoverOptions<T_ItemType extends ItemType> = {
+  onSlow?: () => any;
+  name: string;
+  type: T_ItemType;
+  frameDuration?: number;
+  mover: StateNameProperty<T_ItemType> & string;
+  autoRerun?: boolean; // if it should automatically start the next frame, otherwise it will run when elapsedTime changes
+};
