+++ tsconfig.json
@@ -8,7 +8,8 @@
     "jsx": "react",
     "sourceMap": false,
     "rootDir": "src",
-    "moduleResolution": "node"
+    "moduleResolution": "node",
+    "skipLibCheck": true
   },
   "include": ["src"],
   "exclude": ["node_modules", "dist"]
